requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
email-validator>=1.1.3
psycopg2-binary>=2.9.1
redis>=3.5.3
python-multipart>=0.0.5
python-dateutil==2.9.0 
pyjwt==2.10.1
bcrypt==4.1.1
router>=0.0.1